refactor: mutualize code for konnector icons :art:
